Start work on the hardware simulator
Complete parsing of the main arguments and options, create a few stub classes with little to no functionality.

This early work is going to be expanded upon, and currently the only instructions and registers implemented are for Register-RISC architecture.

We might have missed a lot of things, this is just the base to grow stuff on.