More work on the simulator, interface, loading the code into memory, executing instructions one by one
Interface: created some basic boxes and tooltips, there still is a lot of work with this one, currently just shows the next instruction and is supposed to execute it by button press. Does not work ideally and is due to change, also, launching curses and not ending the program execution normally might (and will) damage your terminal.

Started work on instructions decoding, added some of the first test instructions, the work is supposed to continue in a separate module(s). We are going to split most of this module between files as well, it's just too huge now and is going to be even bigger later.